+++ src/ex_getln.c
@@ -12,6 +12,7 @@
  */
 
 #include "vim.h"
+#include "os/os.h"
 
 /*
  * Variables shared between getcmdline(), redrawcmdline() and others.
