+++ src/os_unix.c
@@ -1244,127 +1244,6 @@ static char * strerror(int err)
 }
 #endif
 
-/*
- * Get absolute file name into "buf[len]".
- *
- * return FAIL for failure, OK for success
- */
-int mch_FullName(
-        char_u      *fname,
-        char_u      *buf,
-        int len,
-        int force                       /* also expand when already absolute path */
-        )
-{
-  int l;
-#ifdef HAVE_FCHDIR
-  int fd = -1;
-  static int dont_fchdir = FALSE;       /* TRUE when fchdir() doesn't work */
-#endif
-  char_u olddir[MAXPATHL];
-  char_u      *p;
-  int retval = OK;
-
-
-
-  /* expand it if forced or not an absolute path */
-  if (force || !mch_isFullName(fname)) {
-    /*
-     * If the file name has a path, change to that directory for a moment,
-     * and then do the getwd() (and get back to where we were).
-     * This will get the correct path name with "../" things.
-     */
-    if ((p = vim_strrchr(fname, '/')) != NULL) {
-#ifdef HAVE_FCHDIR
-      /*
-       * Use fchdir() if possible, it's said to be faster and more
-       * reliable.  But on SunOS 4 it might not work.  Check this by
-       * doing a fchdir() right now.
-       */
-      if (!dont_fchdir) {
-        fd = open(".", O_RDONLY | O_EXTRA, 0);
-        if (fd >= 0 && fchdir(fd) < 0) {
-          close(fd);
-          fd = -1;
-          dont_fchdir = TRUE;               /* don't try again */
-        }
-      }
-#endif
-
-      /* Only change directory when we are sure we can return to where
-       * we are now.  After doing "su" chdir(".") might not work. */
-      if (
-#ifdef HAVE_FCHDIR
-        fd < 0 &&
-#endif
-        (mch_dirname(olddir, MAXPATHL) == FAIL
-         || mch_chdir((char *)olddir) != 0)) {
-        p = NULL;               /* can't get current dir: don't chdir */
-        retval = FAIL;
-      } else   {
-        /* The directory is copied into buf[], to be able to remove
-         * the file name without changing it (could be a string in
-         * read-only memory) */
-        if (p - fname >= len)
-          retval = FAIL;
-        else {
-          vim_strncpy(buf, fname, p - fname);
-          if (mch_chdir((char *)buf))
-            retval = FAIL;
-          else
-            fname = p + 1;
-          *buf = NUL;
-        }
-      }
-    }
-    if (mch_dirname(buf, len) == FAIL) {
-      retval = FAIL;
-      *buf = NUL;
-    }
-    if (p != NULL) {
-#ifdef HAVE_FCHDIR
-      if (fd >= 0) {
-        if (p_verbose >= 5) {
-          verbose_enter();
-          MSG("fchdir() to previous dir");
-          verbose_leave();
-        }
-        l = fchdir(fd);
-        close(fd);
-      } else
-#endif
-      l = mch_chdir((char *)olddir);
-      if (l != 0)
-        EMSG(_(e_prev_dir));
-    }
-
-    l = STRLEN(buf);
-    if (l >= len - 1)
-      retval = FAIL;       /* no space for trailing "/" */
-    else if (l > 0 && buf[l - 1] != '/' && *fname != NUL
-             && STRCMP(fname, ".") != 0)
-      STRCAT(buf, "/");
-  }
-
-  /* Catch file names which are too long. */
-  if (retval == FAIL || (int)(STRLEN(buf) + STRLEN(fname)) >= len)
-    return FAIL;
-
-  /* Do not append ".", "/dir/." is equal to "/dir". */
-  if (STRCMP(fname, ".") != 0)
-    STRCAT(buf, fname);
-
-  return OK;
-}
-
-/*
- * Return TRUE if "fname" does not depend on the current directory.
- */
-int mch_isFullName(char_u *fname)
-{
-  return *fname == '/' || *fname == '~';
-}
-
 #if defined(USE_FNAME_CASE) || defined(PROTO)
 /*
  * Set the case of the file name, if it already exists.  This will cause the
